requirements/requirements.txt
beautifulsoup4==4.12.3
enlighten==1.12.4
jinxed==1.3.0  # TODO: remove when enlighten uses this version
markdown==3.7
platformdirs==4.3.3
puremagic==1.27
pypandoc_binary==1.13
python-frontmatter==1.1.0
pytodotxt==1.5.0
pyyaml==6.0.2
rich==13.8.1
stage-left==0.2.2
vobject==0.9.7
